Revert moving team validation files (they were moved unintentionally)
