+++ api/src/main/java/com/ford/labs/retroquest/team/validation/PasswordConstraint.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package com.ford.labs.retroquest.team;
+package com.ford.labs.retroquest.team.validation;
 
 
 import javax.validation.Constraint;
