+++ api/src/main/java/com/ford/labs/retroquest/team/validation/TeamNameValidator.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package com.ford.labs.retroquest.team;
+package com.ford.labs.retroquest.team.validation;
 
 import com.ford.labs.retroquest.exception.EmptyTeamNameException;
 import com.ford.labs.retroquest.exception.SpecialCharacterTeamNameException;
