+++ api/src/test/java/com/ford/labs/retroquest/team/validation/TeamNameValidatorTest.java
@@ -20,7 +20,6 @@ package com.ford.labs.retroquest.team.validation;
 
 import com.ford.labs.retroquest.exception.EmptyTeamNameException;
 import com.ford.labs.retroquest.exception.SpecialCharacterTeamNameException;
-import com.ford.labs.retroquest.team.TeamNameValidator;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
